Split Formatting from TextUi class into Formatter.java
